// pc/src/app/types/common.type.ts
export interface ResponseBody {
    data: any;
    msg: string;
    status: number;
}
// ...
export interface NavGameInfo {
    gameCode: string;
    gameId: number | string;
    hot: number;
    id: number;
    imgs: any;
    menuNameCn: string;
    menuNameEn: string;
    menuType: string;
    parentId: number;
    parentName: string;
    remark: string;
    sortNo: number;
    subMenus: any;
    template: string;
    url: string;
    link?: string;
    hasSub: boolean;
    totalSub: number;
}
// ...
export interface LoadParams {
    gameCode: string;
    gameId: number | string;
    terminal?: string;
}